Verify against the real releases, and fix what that turned up
`tests/real_release.rs` runs the engine against the actual published
artifacts. Ignored by default -- `cargo test` should not fetch 210 MB on
every run -- because the fixture tests can only prove the pipeline is
self-consistent. They cannot prove sources.toml still describes the ten
releases correctly, and that is the part most likely to rot.

Result: all ten install, hash-verified, with every declared entry point
present. ForgePact's tree is checked in detail against what
build_release.py documents -- ForgePact.exe beside modfiles/ with
AurieCore, YYToolkit, BloodPactPlugin and AuriePatcher.

Running `npm start` found a real one. Vite's watcher recurses into
`src-tauri/target`, where cargo is writing and holding open the
executable it just built; on Windows that is `EBUSY: resource busy or
locked, watch ...hero_siege_toolkit_hub.exe`, and it takes the dev
server down in the middle of the rebuild that caused it. The watcher now
ignores `src-tauri`.

The app itself builds and runs: the window opens, draws its own chrome
and shows the first-run screen.

Clippy is clean. Of the five it found, one was worth more than style:
`SW_SHOWNORMAL as i32` cast a value that is already i32 in
windows-sys 0.59, which would have quietly stopped compiling if the
type ever changed underneath it.

Full suite: 65 Python, 73 Rust unit, 7 end-to-end, 2 ignored real-release.

+++ hub/src-tauri/src/launch.rs
@@ -166,7 +166,7 @@ fn spawn_elevated(entry: &Path, args: &[String], cwd: &Path) -> Result<u32> {
     info.lpVerb = verb.as_ptr();
     info.lpFile = file.as_ptr();
     info.lpDirectory = directory.as_ptr();
-    info.nShow = SW_SHOWNORMAL as i32;
+    info.nShow = SW_SHOWNORMAL;
     if let Some(parameters) = parameters.as_ref() {
         info.lpParameters = parameters.as_ptr();
     }
@@ -242,7 +242,7 @@ fn open_document(path: &Path) -> Result<()> {
             file.as_ptr(),
             std::ptr::null(),
             std::ptr::null(),
-            SW_SHOWNORMAL as i32,
+            SW_SHOWNORMAL,
         )
     };
     // ShellExecuteW returns a value greater than 32 on success. The API is old
